issue #13 Fix :when modifier
Previous to this, if [1 2] and [1 2] were passed as events, this would
not trigger an error because the :when modifier was, naively, checking
repeated events because we would not want to compare them.  This is
because the `for` is going to create repeats.  The issue with this is
that a user would genuinely have add two events that evenly overlap.  To
fix this I use a data structure that has a UUID and I just compare the
UUIDs.