04-clock - remove duplicated functions
There was a lot of duplication throughout the code, so I removed it and
also resolved the issue where when the app loaded, all the hands would
not be properly set until after a certain point.  This is a set further
which a dev could go and it may not be worth it, but it was a good
exercise in reusability.