Move Drumkit to figwheel driven browser app
Prior to this commit we had 3 different apps.  They all did the same
thing, but the only difference was the tooling associated to them.  Now
I see that this is more confusing then anything else, so I am
standardizing on the figwheel package.  The next thing introduce is we
bumped all versions of all dependencies and standardized on a way of
starting the app